Moved set functions to global config
